CLI refactor
