+++ index.js
@@ -1,11 +1,13 @@
 #!/usr/bin/env node
 
-require("./src/ImageSequencer");
+require('./src/ImageSequencer');
 sequencer = ImageSequencer({ ui: false });
-var Spinner = require("ora");
+var fs = require('fs')
+var program = require('commander');
 
-var program = require("commander");
-var readlineSync = require("readline-sync");
+var saveSequence = require('./src/cli/saveSequence.js')
+var installModule = require('./src/cli/installModule.js')
+var sequencerSteps = require('./src/cli/sequencerSteps.js')
 
 function exit(message) {
   console.error(message);
@@ -23,24 +25,11 @@ program
   .option('--install-module [string]', "Module name space seaprated npm package name")
   .parse(process.argv);
 
-if (program.saveSequence) {
-  var params = program.saveSequence.split(' ');
-  sequencer.saveSequence(params[0], params[1]);
-  console.log("\x1b[32m", "Your sequence was saved successfully!!");
-} else if (program.installModule) {
-  console.log(
-    "\x1b[33m%s\x1b[0m",
-    "Please wait while your Module is being Installed...\nThis may take a while!"
-  );
-  
-  var params = program.installModule.split(' ');
-  var spinner = Spinner("Now Installing...").start();
-  require('child_process').execSync(`npm i ${params[1]}`)
-  sequencer.saveNewModule(params[0], params[1]);
-  sequencer.loadNewModule(params[0], require(params[1]));
-  spinner.stop();
-  console.log("\x1b[32m%s\x1b[0m", "Your module was installed successfully!!");
-} else {
+if (program.saveSequence) saveSequence(program, sequencer)
+
+else if (program.installModule) installModule(program, sequencer)
+
+else {
   // Parse step into an array to allow for multiple steps.
   if (!program.step) exit("No steps passed");
   program.step = program.step.split(" ");
@@ -49,12 +38,12 @@ if (program.saveSequence) {
   if (!program.image) exit("Can't read file.");
 
   // User must input an image.
-  require("fs").access(program.image, function(err) {
+  fs.access(program.image, function(err) {
     if (err) exit("Can't read file.");
   });
 
   // User must input a step. If steps exist, check that every step is a valid step.
-  if (!program.step || !validateSteps(program.step))
+  if (!program.step || !(require('./src/CliUtils').validateSteps(program.step, sequencer)))
     exit("Please ensure all steps are valid.");
 
   // If there's no user defined output directory, select a default directory.
@@ -88,124 +77,9 @@ if (program.saveSequence) {
     else {
       outputFilename = null;
     }
-    require("./src/CliUtils").makedir(program.output, () => {
-      console.log('Files will be exported to "' + program.output + '"');
 
-      if (program.basic)
-        console.log("Basic mode is enabled, outputting only final image");
+    sequencerSteps(program, sequencer, outputFilename)
 
-      // Iterate through the steps and retrieve their inputs.
-      program.step.forEach(function(step) {
-        var options = Object.assign({}, sequencer.modulesInfo(step).inputs);
-
-        // If inputs exists, print to console.
-        if (Object.keys(options).length) {
-          console.log("[" + step + "]: Inputs");
-        }
-
-        // If inputs exists, print them out with descriptions.
-        Object.keys(options).forEach(function(input) {
-          // The array below creates a variable number of spaces. This is done with (length + 1).
-          // The extra 4 that makes it (length + 5) is to account for the []: characters
-          console.log(
-            new Array(step.length + 5).join(" ") +
-            input +
-            ": " +
-            options[input].desc
-          );
-        });
-
-        if (program.config) {
-          try {
-            var config = JSON.parse(program.config);
-            console.log(`The parsed options object: `, config);
-          } catch (e) {
-            console.error(
-              "\x1b[31m%s\x1b[0m",
-              `Options(Config) is not a not valid JSON Fallback activate`
-            );
-            program.config = false;
-            console.log(e);
-          }
-        }
-        if (program.config && validateConfig(config, options)) {
-          console.log("Now using Options object");
-          Object.keys(options).forEach(function(input) {
-            options[input] = config[input];
-          });
-        } else {
-          // If inputs exist, iterate through them and prompt for values.
-          Object.keys(options).forEach(function(input) {
-            var value = readlineSync.question(
-              "[" +
-              step +
-              "]: Enter a value for " +
-              input +
-              " (" +
-              options[input].type +
-              ", default: " +
-              options[input].default +
-              "): "
-            );
-            options[input] = value;
-          });
-        }
-        // Add the step and its inputs to the sequencer.
-        sequencer.addSteps(step, options);
-      });
-
-      var spinnerObj = { succeed: () => { }, stop: () => { } };
-      if (!process.env.TEST)
-        spinnerObj = Spinner("Your Image is being processed..").start();
-
-      // Run the sequencer.
-      sequencer.run({ progressObj: spinnerObj }, function() {
-        // Export all images or final image as binary files.
-        sequencer.exportBin(program.output, program.basic, outputFilename);
-
-        //check if spinner was not overriden stop it
-        if (!spinnerObj.overrideFlag) {
-          spinnerObj.succeed();
-          console.log(`\nDone!!`);
-        }
-      });
-    });
   });
 
-  // Takes an array of steps and checks if they are valid steps for the sequencer.
-  function validateSteps(steps) {
-    // Assume all are valid in the beginning.
-    var valid = true;
-    steps.forEach(function(step) {
-      // If any step in the array is not valid (not a property of modulesInfo), set valid to false.
-      if (!sequencer.modulesInfo().hasOwnProperty(step)) {
-        valid = false;
-      }
-    });
-
-    // Return valid. (If all of the steps are valid properties, valid will have remained true).
-    return valid;
-  }
-
-  //Takes config and options object and checks if all the keys exist in config
-  function validateConfig(config_, options_) {
-    options_ = Object.keys(options_);
-    if (
-      (function() {
-        for (var input in options_) {
-          if (!config_[options_[input]]) {
-            console.error(
-              "\x1b[31m%s\x1b[0m",
-              `Options Object does not have the required details "${
-              options_[input]
-              }" not specified. Fallback case activated`
-            );
-            return false;
-          }
-        }
-      })() == false
-    )
-      return false;
-    else return true;
-  }
 }
